requirements.txt
python-telegram-bot
Flask
matplotlib